Handle renaming column casing
products -> Products and the like

need to worry about one issue though:
what if they _add_ a new column that has a name that is the same but
for casing:
products and Products. We would probably identify arbitrarily one of
these with the existing products column and probably bad things will
happen. Need to extend this check that identifies the casing
indifference only in the case that there is no ambiguity